requirements.txt
wheel
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
cachetools==5.1.0
celery==5.2.6
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake~=3.22.4
Deprecated==1.2.13
dlib==19.24.0
Flask==2.1.2
Flask-PyMongo==2.3.0
Flask-Mail==0.9.1
gunicorn==20.1.0
idna==3.3
imageio==2.19.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
nanoid==2.0.0
networkx==2.8.2
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pymongo==4.1.1
pyparsing==3.0.9
pytz==2022.1
PyWavelets==1.3.0
redis==4.3.1
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.1
six==1.16.0
SQLAlchemy==1.4.36
tifffile==2022.5.4
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
python-dotenv~=0.21.0
bcrypt~=4.0.1
aiohttp~=3.8.3